update: added AttackComponent for PlayerEntity; click on Gold in ClickSystem; EntityFactory(renamed) - added new funs for different entities + added attr for them; EntityTypeComponent added new types and removed OBSTACLE(no need); GoldComponent now has value; HealthComponent replaced int with long; SpawnSystem now spawns gold with enemy/boss death

+++ core/src/main/kotlin/com/neymeha/game/seeinggod/ecs/components/HealthComponent.kt
@@ -11,6 +11,6 @@ import com.badlogic.ashley.core.Component
  * @property currentHealth Текущее количество здоровья сущности.
  */
 class HealthComponent(
-    var maxHealth: Int,    // Максимальное количество здоровья сущности.
-    var currentHealth: Int // Текущее количество здоровья сущности.
+    var maxHealth: Long,    // Максимальное количество здоровья сущности.
+    var currentHealth: Long // Текущее количество здоровья сущности.
 ) : Component
